google_adk/requirements.txt
google-adk
google-genai
nest-asyncio
python-dotenv